requirements.txt
torch==1.1.0
transformer==2.2.2
seqeval>=0.0.12
pytorch-crf==0.7.2